+++ components/elements/accordion/Accordion.tsx
@@ -5,6 +5,8 @@ import * as AccordionPrimitive from "@radix-ui/react-accordion";
 import { Chip, ChipTypes } from "@elements/index";
 
 import { cn } from "../../util";
+import { AccordionContent } from "./AccordionContent";
+import { AccordionTrigger } from "./AccordionTrigger";
 
 export type AccordionItemProps = {
   trigger: string;
@@ -45,7 +47,7 @@ const Accordion = React.forwardRef<
         })}
       >
         {items.map((item, index) => (
-          <AccordionItem
+          <AccordionPrimitive.Item
             disabled={item.disabled || false}
             className={cn(itemClassNames, "hawa-rounded")}
             key={index}
@@ -89,97 +91,13 @@ const Accordion = React.forwardRef<
             >
               {item.content}
             </AccordionContent>
-          </AccordionItem>
+          </AccordionPrimitive.Item>
         ))}
       </div>
     </AccordionPrimitive.Root>
   )
 );
 
-const AccordionItem = React.forwardRef<
-  React.ElementRef<typeof AccordionPrimitive.Item>,
-  React.ComponentPropsWithoutRef<typeof AccordionPrimitive.Item>
->(({ className, ...props }, ref) => (
-  <AccordionPrimitive.Item ref={ref} className={cn(className)} {...props} />
-));
-
-const AccordionTrigger = React.forwardRef<
-  React.ElementRef<typeof AccordionPrimitive.Trigger>,
-  React.ComponentPropsWithoutRef<typeof AccordionPrimitive.Trigger> & {
-    unstyled?: boolean;
-    hideArrow?: boolean;
-  }
->(({ className, children, ...props }, ref) => (
-  <AccordionPrimitive.Header className="hawa-flex">
-    <AccordionPrimitive.Trigger
-      ref={ref}
-      className={
-        !props.unstyled
-          ? cn(
-              "hawa-flex hawa-flex-1 hawa-items-center hawa-justify-between hawa-bg-muted/50 hawa-p-4 hawa-font-medium hawa-transition-all [&[data-state=open]>svg]:hawa-rotate-180",
-              props.disabled
-                ? "hawa-bg-muted/30 hawa-text-muted-foreground/50"
-                : "hover:hawa-bg-muted",
-              className
-            )
-          : className
-      }
-      {...props}
-    >
-      {children}
-      {!props.hideArrow && (
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          className="hawa-icon hawa-shrink-0 hawa-transition-transform hawa-duration-200"
-        >
-          <path d="m6 9 6 6 6-6" />
-        </svg>
-      )}
-    </AccordionPrimitive.Trigger>
-  </AccordionPrimitive.Header>
-));
-
-const AccordionContent = React.forwardRef<
-  React.ElementRef<typeof AccordionPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof AccordionPrimitive.Content> & {
-    unstyled?: boolean;
-  }
->(({ className, children, ...props }, ref) => (
-  <AccordionPrimitive.Content
-    ref={ref}
-    className={
-      !props.unstyled
-        ? cn(
-            "hawa-overflow-hidden hawa-text-sm hawa-transition-all data-[state=closed]:hawa-animate-accordion-up data-[state=open]:hawa-animate-accordion-down",
-            className
-          )
-        : className
-    }
-    {...props}
-  >
-    <div className="hawa-bg-background hawa-p-4">{children}</div>
-  </AccordionPrimitive.Content>
-));
-
-const AccordionRoot = AccordionPrimitive.Root;
-
 Accordion.displayName = "Accordion";
-AccordionItem.displayName = "AccordionItem";
-AccordionTrigger.displayName = AccordionPrimitive.Trigger.displayName;
-AccordionContent.displayName = AccordionPrimitive.Content.displayName;
 
-export {
-  Accordion,
-  AccordionItem,
-  AccordionRoot,
-  AccordionTrigger,
-  AccordionContent
-};
+export { Accordion };