+++ components/elements/dropdownMenu/DropdownMenu.tsx
@@ -2,8 +2,8 @@ import * as React from "react";
 
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
 
-import { DirectionType } from "../types/commonTypes";
-import { cn } from "../util";
+import { DirectionType } from "../../types/commonTypes";
+import { cn } from "../../util";
 
 const DropdownMenuRoot = DropdownMenuPrimitive.Root;
 const DropdownMenuTrigger = DropdownMenuPrimitive.Trigger;