+++ components/elements/loading/Loading.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from "react";
 
-import { cn } from "../util";
+import { cn } from "../../util";
 
 type LoadingTypes = {
   /** Specifies the size of the loading component.*/